theem file can thiet

+++ src/main/java/com/project/shopapp/services/impl/DriveService.java
@@ -3,16 +3,26 @@ package com.project.shopapp.services.impl;
 import com.google.api.client.googleapis.auth.oauth2.GoogleCredential;
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
 import com.google.api.client.http.FileContent;
+import com.google.api.client.http.HttpRequestInitializer;
 import com.google.api.client.json.JsonFactory;
 import com.google.api.client.json.gson.GsonFactory;
 import com.google.api.services.drive.Drive;
 import com.google.api.services.drive.DriveScopes;
+import com.google.auth.http.HttpCredentialsAdapter;
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.auth.oauth2.ServiceAccountCredentials;
 import com.project.shopapp.responses.DriveResponse;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import com.google.auth.oauth2.GoogleCredentials;
+import com.google.api.services.drive.Drive;
+import com.google.api.services.drive.DriveScopes;
 
+import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.security.GeneralSecurityException;
@@ -22,10 +32,8 @@ import java.util.Collections;
 public class DriveService {
     private static final JsonFactory JSON_FACTORY = GsonFactory.getDefaultInstance();
     //private static final String SERVICE_ACOUNT_KEY_PATH = getPathToGoodleCredentials();
-
-    private static final String SERVICE_ACOUNT_KEY_PATH = "src/main/resources/cred.json";
-
-
+    @Value("${google-drive.folder-id}")
+    private String folderId;
 
     private static String getPathToGoodleCredentials() {
         String currentDirectory = System.getProperty("user.dir");
@@ -48,8 +56,12 @@ public class DriveService {
                     .setFields("id").execute();
             String imageUrl = "https://drive.google.com/uc?export=view&id="+uploadedFile.getId();
             System.out.println("IMAGE URL: " + imageUrl);
-            file.delete();
+
+//            file.delete();
             res.setStatus(200);
+            if (res.getStatus() == 200) {
+                file.delete();
+            }
             res.setMessage("Image Successfully Uploaded To Drive");
             res.setUrl(imageUrl);
         }catch (Exception e){
@@ -77,4 +89,5 @@ public class DriveService {
                 .build();
 
     }
+
 }
